requirements.txt
streamlit==0.64.0
pandas==1.1.0
numpy==1.19.1
scikit-learn==0.23.2
nltk==3.5
PyPDF2==1.26.0
sklearn==0.0
hyperlink==20.0.1
mysql-connector-python==8.0.2
